Create ICO
http://www.icohelper.info/index.php?p=home
Icohelper is a project of juvemobile.it and of its team that is born from the need of making a team leader for the development of icos for other users. Aim:
Our aim is to help the next ico creator and guide the users of our own Zico to invest on his too, helping it grow. To receive our help in creating an Ico, the users will have to buy our coins and pay us with them, so that these transactions will help us raise the value of our Icohelper Coin.